Cut fixes

Closes #1634 and #1636

See merge request !599

+++ src/scip/pub_misc.h
@@ -488,9 +488,8 @@ INLINE static
 uint32_t SCIPrealHashCode(double x)
 {
    int exp;
-   return (((uint32_t)(uint16_t)(int16_t)ldexp(frexp(x, &exp), 15))<<16) | (uint32_t)exp;
+   return (((uint32_t)(uint16_t)(int16_t)ldexp(frexp(x, &exp), 15))<<16) | (uint32_t)(uint16_t)exp;
 }
-
 
 /** creates a hash table */
 EXTERN
